Install RegexConverter on all handlers
I'll need this in order to port the district detail url handlers, and it
just seems like a good idea for us to have available everywhere
