+++ src/backend/web/main.py
@@ -4,6 +4,7 @@ from flask_wtf.csrf import CSRFProtect
 from backend.common.flask_cache import configure_flask_cache
 from backend.common.logging import configure_logging
 from backend.common.middleware import install_middleware
+from backend.common.url_converters import install_url_converters
 from backend.web.auth import _user_context_processor
 from backend.web.handlers.account import blueprint as account_blueprint
 from backend.web.handlers.error import handle_404, handle_500
@@ -30,6 +31,7 @@ configure_logging()
 
 app = Flask(__name__)
 install_middleware(app)
+install_url_converters(app)
 configure_flask_cache(app)
 
 csrf = CSRFProtect()
